pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentfier"
version = "0.1.0"
description = "Agentfier - Analyze codebases and generate agent-native conversion specifications"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "anthropic>=0.40.0",
    "streamlit>=1.40.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "gitpython>=3.1.40",
    "httpx>=0.27.0",
    "graphviz>=0.20.0",
    "chardet>=5.2.0",
    "defusedxml>=0.7.1",
    "toml>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]
